Refactor: move Chat to it own space
