+++ src/components/Chat/Request/RequestChat.tsx
@@ -1,14 +1,16 @@
-import { Stack } from "@mui/material";
+import { Stack, Typography } from "@mui/material";
 import RequestMessageType from "../../../Interface/Message/RequestMessageType";
 import { useSmoothScrollToBottom } from "../../../utils/chat";
-import RequestMessage from "../../Messages/Request/RequestMessage";
+import { BasicRequestMessage } from "./BasicRequestMessage";
 
 export function RequestChat({ message }: { message: RequestMessageType }) {
 	const divRef = useSmoothScrollToBottom();
 
 	return (
 		<Stack ref={divRef} width={"100%"} alignItems={"end"}>
-			<RequestMessage message={message} />
+			<BasicRequestMessage>
+				<Typography sx={{ textAlign: "left" }}>{message.message}</Typography>
+			</BasicRequestMessage>
 		</Stack>
 	);
 }
