+++ src/components/Conversation/index.tsx
@@ -1,9 +1,9 @@
 import { Stack } from "@mui/material";
 import useMessageStatus from "../../hooks/Request/useMessageStatus";
 import { useAppSelector } from "../../store/store";
-import ChatChunk from "./Chat/ChatChunk";
-import { LoadingResponseMessage } from "./Chat/LoadingResponseMessage";
-import { ResponseChat } from "./Response/ResponseChat";
+import ChatChunk from "../Chat/ChatChunk";
+import { LoadingResponseMessage } from "../Chat/LoadingResponseMessage";
+import { ResponseChat } from "../Chat/Response/ResponseChat";
 
 export default function Conversation() {
 	const { status: isLoading } = useMessageStatus();
